Added Match and tested match, there are a couple bugs, future improvements could be fixing up the dialogue and showing who goes first and what the player's balance is

+++ src/Character.java
@@ -24,12 +24,12 @@ public class Character {
     }
 
     //add aura
-    public void addAura(float amount){
+    public void addAura(double amount){
         aura += amount;
     }
 
     //remove aura
-    public void loseAura(float amount){
+    public void loseAura(double amount){
         aura -= amount;
     }
